+++ src/types/components/event.ts
@@ -5,7 +5,7 @@ import { RRule } from 'rrule'
 type Event = {
     dtStamp: string | ComplexDate
     uid: string
-    start?: string
+    start?: string | ComplexDate
     class?: Classification
     created?: string
     description?: string