+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 - Removed travis ci
 
+### Fixed
+
+- Fixed event start to support ComplexDate
+
 ## [0.3.2] - 2020-04-21
 
 ### Fixed